requirements.txt
python-dotenv>=1.0.0
pandas>=1.5.0
pinecone-client>=3.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
